fix: PR fixes and fix to home page button not closing dialog
