+++ src/components/FullscreenDialog/index.ts
@@ -1,0 +1,1 @@
+export { default as FullscreenDialog } from './FullscreenDialog'
