+++ src/components/NewsletterSubscriptionForm/index.ts
@@ -1,0 +1,1 @@
+export { default as NewsletterSubscriptionForm } from './NewsletterSubscriptionForm'
