--- src/components/SubscribeDialogue/index.ts
@@ -1,1 +1,0 @@
-export { default as SubscribeDialogue } from './SubscribeDialogue'
