+++ src/containers/NewsletterSubscriptionDialog/index.ts
@@ -1,0 +1,1 @@
+export { default as NewsletterSubscriptionDialog } from './NewsletterSubscriptionDialog'
